# docs/openapi.yaml
openapi: 3.0.3
info:
  title: BookVerse Recommendations Service API
  description: |
    ## Overview
    
    The BookVerse Recommendations Service provides intelligent book recommendations using advanced machine learning algorithms. The service delivers personalized recommendations based on user behavior, book content analysis, and collaborative filtering techniques.
    
    ## Key Features
    
    - **Personalized Recommendations**: ML-driven suggestions based on user preferences
    - **Content-Based Filtering**: Recommendations based on book content similarity
    - **Collaborative Filtering**: User behavior-based recommendation engine
    - **Real-Time Processing**: Instant recommendation generation and updates
    - **A/B Testing Support**: Multiple recommendation algorithms for testing
    - **Analytics Integration**: Comprehensive tracking of recommendation performance
    
    ## Machine Learning Models
    
    - **Content Similarity**: TF-IDF and cosine similarity for content matching
    - **Collaborative Filtering**: Matrix factorization and neighborhood methods
    - **Deep Learning**: Neural networks for complex pattern recognition
    - **Ensemble Methods**: Combining multiple algorithms for optimal results
    
    ## Performance
    
    - **Response Time**: < 100ms for recommendation requests
    - **Throughput**: 10,000+ recommendations per second
    - **Model Updates**: Real-time learning from user interactions
...
    
  version: 1.1.0
  contact:
...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /recommendations/trending:
    get:
      summary: Get trending book recommendations
      description: |
        Retrieve currently trending books based on real-time analytics including sales velocity, user engagement, ratings, and social media mentions.
        
        **Trending Factors:**
        - Sales velocity and growth rate
        - User ratings and review activity
        - Search frequency and click-through rates
        - Social media mentions and sharing
        - Seasonal and temporal patterns
        
        **Time Windows:**
        - **hourly**: Last 6 hours (real-time trends)
        - **daily**: Last 24 hours (current trends)
        - **weekly**: Last 7 days (popular trends)
        - **monthly**: Last 30 days (sustained trends)
        
        **Categories:**
        - Global trending across all categories
        - Category-specific trending books
        - New releases gaining traction
        - Rising stars with high growth rates
      operationId: getTrendingRecommendations
      tags:
        - Recommendations
        - Trending
      parameters:
        - name: time_window
          in: query
          description: Time window for trending analysis
          schema:
            type: string
            enum: [hourly, daily, weekly, monthly]
            default: daily
            example: daily
        
        - name: category
          in: query
          description: Specific category for trending books
          schema:
            type: string
            enum: [programming, fiction, non-fiction, science, technology, business, education, reference]
            example: programming
        
        - name: trend_type
          in: query
          description: Type of trending analysis
          schema:
            type: string
            enum: [hot, rising, new, sustained]
            default: hot
            example: hot
        
        - name: region
          in: query
          description: Geographic region for localized trends
          schema:
            type: string
            enum: [global, north_america, europe, asia, other]
            default: global
            example: global
        
        - name: limit
          in: query
          description: Maximum number of trending books to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
            example: 20

      responses:
        '200':
          description: Trending recommendations retrieved successfully
          headers:
            X-Trend-Analysis-Time:
              description: Timestamp of trend analysis
              schema:
                type: string
                format: date-time
                example: "2024-01-25T16:00:00Z"
            X-Data-Freshness-Minutes:
              description: Age of trending data in minutes
              schema:
                type: integer
                example: 5
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/TrendingRecommendation'
                  trend_metadata:
                    $ref: '#/components/schemas/TrendMetadata'
              examples:
                daily_trending:
                  summary: Daily trending books
                  value:
                    data:
                      - book_id: 567
                        title: "AI and Machine Learning Fundamentals"
                        author: "Sarah Johnson"
                        rank: 1
                        trend_score: 0.98
                        velocity: 2.34
                        growth_rate: 0.45
                        category: "technology"
                        reasons: ["high_sales_velocity", "positive_reviews", "social_mentions"]
                        metrics:
                          sales_increase_percent: 234
                          rating_trend: 0.2
                          search_volume_change: 1.8
                          social_mentions: 127
                    trend_metadata:
                      analysis_time: "2024-01-25T16:00:00Z"
                      time_window: "daily"
                      total_analyzed: 15420
                      data_freshness_minutes: 5
                      algorithm_version: "v1.3.2"

        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /recommendations/similar:
    get:
      summary: Get books similar to a list of books
      description: |
        Find books similar to a provided list of books using advanced content analysis and collaborative filtering. Ideal for creating "customers who bought these also bought" sections.
        
        **Similarity Analysis:**
        - Content features: genre, themes, writing style, topics
        - Metadata similarity: author, publisher, publication era
        - User behavior patterns: co-purchasing, co-rating
        - Advanced NLP: semantic similarity of descriptions
        
        **Use Cases:**
        - Shopping cart cross-selling recommendations
        - "Similar books" sections on product pages
        - Reading list expansion suggestions
        - Collection building recommendations
      operationId: getSimilarBooks
      tags:
        - Recommendations
        - Similarity
      parameters:
        - name: book_ids
          in: query
          required: true
          description: |
            Comma-separated list of book IDs to find similar books for.
            Maximum 10 books per request.
          schema:
            type: string
            pattern: '^[0-9]+(,[0-9]+)*$'
            example: "123,456,789"
        
        - name: similarity_type
          in: query
          description: Type of similarity analysis to perform
          schema:
            type: string
            enum: [content, behavioral, hybrid]
            default: hybrid
            example: hybrid
        
        - name: exclude_input
          in: query
          description: Exclude the input books from recommendations
          schema:
            type: boolean
            default: true
            example: true
        
        - name: min_similarity
          in: query
          description: Minimum similarity threshold (0-1)
          schema:
            type: number
            format: float
            minimum: 0
            maximum: 1
            default: 0.3
            example: 0.5

      responses:
        '200':
          description: Similar books found successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/SimilarBook'
                  similarity_analysis:
                    $ref: '#/components/schemas/SimilarityAnalysis'
              examples:
                similar_books_result:
                  summary: Books similar to provided list
                  value:
                    data:
                      - book_id: 890
                        title: "Software Architecture: The Hard Parts"
                        author: "Neal Ford"
                        similarity_score: 0.87
                        similarity_reasons: ["same_genre", "similar_themes", "co_purchased"]
                        common_features: ["software engineering", "architecture", "best practices"]
                    similarity_analysis:
                      input_books: [123, 456, 789]
                      total_candidates: 2847
                      similarity_method: "hybrid"
                      processing_time_ms: 156

        '400':
          description: Invalid book IDs or too many books provided
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: "https://docs.bookverse.com/errors/invalid-book-list"
                title: "Invalid Book List"
                status: 400
                detail: "Maximum 10 books allowed per request. Received 15 books."
                instance: "/recommendations/similar"
        '404':
          description: One or more books not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: "https://docs.bookverse.com/errors/books-not-found"
                title: "Books Not Found"
                status: 404
                detail: "Books not found: [999, 888]"
                instance: "/recommendations/similar"
                missing_books: [999, 888]
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Recommendation:
      type: object
      required:
        - book_id
        - title
        - author
        - confidence_score
        - reason
      properties:
        book_id:
          type: integer
          format: int64
          description: Unique book identifier
          example: 456
        title:
          type: string
          description: Book title
          example: "Design Patterns: Elements of Reusable Object-Oriented Software"
        author:
          type: string
          description: Book author
          example: "Gang of Four"
        isbn:
          type: string
          description: Book ISBN
          example: "978-0201633612"
        price:
          type: number
          format: decimal
          description: Current book price
          example: 54.99
        currency:
          type: string
          enum: [USD, EUR, GBP, CAD]
          description: Price currency
          example: "USD"
        rating:
          type: number
          format: float
          minimum: 0
          maximum: 5
          description: Average user rating
          example: 4.6
        review_count:
          type: integer
          minimum: 0
          description: Number of reviews
          example: 128
        category:
          type: string
          description: Book category
          example: "programming"
        cover_image_url:
          type: string
          format: uri
          description: Book cover image URL
          example: "https://images.bookverse.com/covers/design-patterns.jpg"
        availability:
          type: string
          enum: [available, out_of_stock, discontinued]
          description: Current availability
          example: "available"
        confidence_score:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: Recommendation confidence (0-1)
          example: 0.92
        similarity_score:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: Content similarity score (0-1)
          example: 0.87
        reason:
          type: string
          enum: [content_similarity, collaborative_filtering, popularity, trending, hybrid]
          description: Primary recommendation reason
          example: "content_similarity"
        explanation:
          type: string
          description: Human-readable explanation
          example: "Similar to books you've read about software architecture"
        recommendation_id:
          type: string
          description: Unique recommendation identifier for tracking
          example: "rec_abc123def456"
        position:
          type: integer
          minimum: 1
          description: Position in recommendation list
          example: 1
    TrendingRecommendation:
      allOf:
        - $ref: '#/components/schemas/Recommendation'
        - type: object
          properties:
            rank:
              type: integer
              minimum: 1
              description: Trending rank position
              example: 1
            trend_score:
              type: number
              format: float
              minimum: 0
              maximum: 1
              description: Overall trending score
              example: 0.98
            velocity:
              type: number
              format: float
              description: Trend velocity factor
              example: 2.34
            growth_rate:
              type: number
              format: float
              description: Growth rate percentage
              example: 0.45
            reasons:
              type: array
              items:
                type: string
                enum: [high_sales_velocity, positive_reviews, social_mentions, search_spike, new_release]
              description: Factors contributing to trending status
              example: ["high_sales_velocity", "positive_reviews"]
            metrics:
              type: object
              properties:
                sales_increase_percent:
                  type: number
                  format: float
                  description: Sales increase percentage
                  example: 234
                rating_trend:
                  type: number
                  format: float
                  description: Rating trend direction
                  example: 0.2
                search_volume_change:
                  type: number
                  format: float
                  description: Search volume change factor
                  example: 1.8
                social_mentions:
                  type: integer
                  description: Social media mentions count
                  example: 127
    SimilarBook:
      allOf:
        - $ref: '#/components/schemas/Recommendation'
        - type: object
          properties:
            similarity_reasons:
              type: array
              items:
                type: string
                enum: [same_genre, similar_themes, same_author, co_purchased, co_rated, content_features]
              description: Specific similarity factors
              example: ["same_genre", "similar_themes", "co_purchased"]
            common_features:
              type: array
              items:
                type: string
              description: Common content features
              example: ["software engineering", "architecture", "best practices"]
            related_books:
              type: array
              items:
                type: integer
              description: IDs of books that contributed to similarity
              example: [123, 456]
...
    TrendMetadata:
      type: object
      properties:
        analysis_time:
          type: string
          format: date-time
          description: When trend analysis was performed
          example: "2024-01-25T16:00:00Z"
        time_window:
          type: string
          description: Analysis time window
          example: "daily"
        total_analyzed:
          type: integer
          description: Total books analyzed for trends
          example: 15420
        data_freshness_minutes:
          type: integer
          description: Age of underlying data
          example: 5
        algorithm_version:
          type: string
          description: Trending algorithm version
          example: "v1.3.2"
        confidence_threshold:
          type: number
          format: float
          description: Minimum confidence for trending status
          example: 0.75
    SimilarityAnalysis:
      type: object
      properties:
        input_books:
          type: array
          items:
            type: integer
          description: Input book IDs
          example: [123, 456, 789]
        total_candidates:
          type: integer
          description: Total candidate books analyzed
          example: 2847
        similarity_method:
          type: string
          description: Similarity calculation method
          example: "hybrid"
        processing_time_ms:
          type: integer
          description: Analysis processing time
          example: 156
        similarity_distribution:
          type: object
          properties:
            high_similarity:
              type: integer
              description: Books with similarity > 0.8
              example: 12
            medium_similarity:
              type: integer
              description: Books with similarity 0.5-0.8
              example: 45
            low_similarity:
              type: integer
              description: Books with similarity 0.3-0.5
              example: 78
...
    ProblemDetails:
      type: object
      description: RFC 7807 Problem Details for HTTP APIs
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
          description: URI reference identifying the problem type
          example: "https://docs.bookverse.com/errors/book-not-found"
        title:
          type: string
          description: Short, human-readable summary
          example: "Book Not Found"
        status:
          type: integer
          description: HTTP status code
          example: 404
        detail:
          type: string
          description: Human-readable explanation
          example: "No book found with ID 123"
        instance:
          type: string
          format: uri
          description: URI reference for this specific occurrence
          example: "/recommendations"
  responses:
    BadRequest:
      description: Bad Request - Invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: "https://docs.bookverse.com/errors/validation-error"
            title: "Validation Error"
            status: 400
            detail: "Invalid limit parameter. Must be between 1 and 50."
            instance: "/recommendations"
    Unauthorized:
      description: Unauthorized - Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: "https://docs.bookverse.com/errors/unauthorized"
            title: "Authentication Required"
            status: 401
            detail: "Valid authentication token required"
            instance: "/recommendations"
    RateLimited:
      description: Too Many Requests - Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          description: Request limit per hour
          schema:
            type: integer
            example: 5000
        X-RateLimit-Remaining:
          description: Requests remaining in current window
          schema:
            type: integer
            example: 0
        Retry-After:
          description: Seconds to wait before retrying
          schema:
            type: integer
            example: 3600
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: "https://docs.bookverse.com/errors/rate-limit-exceeded"
            title: "Rate Limit Exceeded"
            status: 429
            detail: "Request rate limit of 5000 per hour exceeded"
            instance: "/recommendations"
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: "https://docs.bookverse.com/errors/internal-server-error"
            title: "Internal Server Error"
            status: 500
            detail: "An unexpected error occurred while processing recommendations"
            instance: "/recommendations"
...